pyproject.toml
[tool.poetry]
name = "yatube"
version = "0.1.0"
description = "Social network of bloggers"
authors = ["Kuznetsov Sergey <monteg179@yandex.ru>"]
readme = "README.md"
packages = [{include = "django"}]

[tool.poetry.dependencies]
python = "^3.11"
django = "^5.0.2"
whitenoise = "^6.6.0"
psycopg2-binary = "^2.9.9"
uvicorn = "^0.28.0"
python-dotenv = "^1.0.1"

[tool.poetry.group.dev.dependencies]
flake8 = "^7.0.0"
isort = "^5.13.2"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
